pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "html2text>=2025.4.15",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-community>=2.0.7",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.3.4",
    "langsmith>=0.4.4",
    "markdown>=3.8.2",
    "python-box==7.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uv==0.7.16",
]
